Update imports after folder renaming
