+++ assets/js/blocks/cart/inner-blocks/index.tsx
@@ -15,5 +15,5 @@ import './cart-order-summary-fee';
 import './cart-order-summary-discount';
 import './cart-order-summary-shipping';
 import './cart-order-summary-coupon-form';
-import './cart-order-summary-taxes';
+import './cart-order-summary-taxes-block';
 import './cart-order-summary-heading';
