+++ bin/webpack-entries.js
@@ -46,7 +46,7 @@ const blocks = {
 	'active-filters': {},
 	cart: {},
 	'cart-order-summary-taxes-block': {
-		customDir: 'cart/inner-blocks/cart-order-summary-taxes',
+		customDir: 'cart/inner-blocks/cart-order-summary-taxes-block',
 	},
 	checkout: {},
 	'mini-cart': {},
